Moved synthesis fixer from transformations -> synthesis
