+++ transformations_test.go
@@ -5,7 +5,6 @@ import (
 	"io/ioutil"
 	"os"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -153,57 +152,6 @@ func TestWriteCodonJSON(t *testing.T) {
 	if diff := cmp.Diff(testCodonTable, readTestCodonTable); diff != "" {
 		t.Errorf(" mismatch (-want +got):\n%s", diff)
 	}
-
-}
-
-/******************************************************************************
-
-Synthesis Fixer tests begin here
-
-Generally, we are taking tests previously used in DnaChisel along with a few of
-our own tests.
-
-https://github.com/Edinburgh-Genome-Foundry/DnaChisel/tree/master/tests/builtin_specifications
-
-The following tests are built into DnaChisel:
-
-- test_AllowPrimer
-- test_AvoidBlastMatches
-- test_AvoidChanges
-- test_AvoidHairpins
-- test_AvoidMatches
-- test_AvoidRareCodons
-- test_AvoidStopCodons
-- test_CodonOptimize
-- test_EnforceChanges
-- test_EnforceChoice
-- test_EnforceGCContent
-- test_EnforcePatternOccurence
-- test_EnforceRegionsCompatibility
-- test_EnforceSequence
-- test_EnforceTerminalGCContent
-- test_EnforceTranslation
-- test_SequenceLengthBounds
-- test_UniquifyAllKmers
-
-
-
-
-******************************************************************************/
-
-func ExampleFixCds() {
-	bla := "ATGAGTATTCAACATTTCCGTGTCGCCCTTATTCCCTTTTTTGCGGCATTTTGCCTTCCTGTTTTTGCTCACCCAGAAACGCTGGTGAAAGTAAAAGATGCTGAAGATCAGTTGGGTGCACGAGTGGGTTACATCGAACTGGATCTCAACAGCGGTAAGATCCTTGAGAGTTTTCGCCCCGAAGAACGTTTTCCAATGATGAGCACTTTTAAAGTTCTGCTATGTGGCGCGGTATTATCCCGTATTGACGCCGGGCAAGAGCAACTCGGTCGCCGCATACACTATTCTCAGAATGACTTGGTTGAGTACTCACCAGTCACAGAAAAGCATCTTACGGATGGCATGACAGTAAGAGAATTATGCAGTGCTGCCATAACCATGAGTGATAACACTGCGGCCAACTTACTTCTGACAACGATCGGAGGACCGAAGGAGCTAACCGCTTTTTTGCACAACATGGGGGATCATGTAACTCGCCTTGATCGTTGGGAACCGGAGCTGAATGAAGCCATACCAAACGACGAGCGTGACACCACGATGCCTGTAGCAATGGCAACAACGTTGCGCAAACTATTAACTGGCGAACTACTTACTCTAGCTTCCCGGCAACAATTAATAGACTGGATGGAGGCGGATAAAGTTGCAGGACCACTTCTGCGCTCGGCCCTTCCGGCTGGCTGGTTTATTGCTGATAAATCTGGAGCCGGTGAGCGTGGGTCTCGCGGTATCATTGCAGCACTGGGGCCAGATGGTAAGCCCTCCCGTATCGTAGTTATCTACACGACGGGGAGTCAGGCAACTATGGATGAACGAAATAGACAGATCGCTGAGATAGGTGCCTCACTGATTAAGCATTGGTAA"
-	sequence := ReadGbk("data/ecoli-mg1655.gff")
-	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
-
-	var functions []func(string, chan DnaSuggestion, *sync.WaitGroup)
-	functions = append(functions, FindBsaI)
-
-	fixedSeq, _ := FixCds(bla, optimizationTable, functions)
-	fmt.Println(fixedSeq)
-
-	// Output: ATGAGTATTCAACATTTCCGTGTCGCCCTTATTCCCTTTTTTGCGGCATTTTGCCTTCCTGTTTTTGCTCACCCAGAAACGCTGGTGAAAGTAAAAGATGCTGAAGATCAGTTGGGTGCACGAGTGGGTTACATCGAACTGGATCTCAACAGCGGTAAGATCCTTGAGAGTTTTCGCCCCGAAGAACGTTTTCCAATGATGAGCACTTTTAAAGTTCTGCTATGTGGCGCGGTATTATCCCGTATTGACGCCGGGCAAGAGCAACTCGGTCGCCGCATACACTATTCTCAGAATGACTTGGTTGAGTACTCACCAGTCACAGAAAAGCATCTTACGGATGGCATGACAGTAAGAGAATTATGCAGTGCTGCCATAACCATGAGTGATAACACTGCGGCCAACTTACTTCTGACAACGATCGGAGGACCGAAGGAGCTAACCGCTTTTTTGCACAACATGGGGGATCATGTAACTCGCCTTGATCGTTGGGAACCGGAGCTGAATGAAGCCATACCAAACGACGAGCGTGACACCACGATGCCTGTAGCAATGGCAACAACGTTGCGCAAACTATTAACTGGCGAACTACTTACTCTAGCTTCCCGGCAACAATTAATAGACTGGATGGAGGCGGATAAAGTTGCAGGACCACTTCTGCGCTCGGCCCTTCCGGCTGGCTGGTTTATTGCTGATAAATCTGGAGCCGGTGAGCGTGGCTCTCGCGGTATCATTGCAGCACTGGGGCCAGATGGTAAGCCCTCCCGTATCGTAGTTATCTACACGACGGGGAGTCAGGCAACTATGGATGAACGAAATAGACAGATCGCTGAGATAGGTGCCTCACTGATTAAGCATTGG
 
 }
